incusd/scriptlet: Refactor log
